fugit::kind

Trait Kind

pub trait Kind: Copy + Sealed {
    const DEBUG_NAME: &'static str;
}

The set of timelines an Instant can be on.

Sealed: the kinds are Monotonic and Wrapping.

Required Associated Constants

const DEBUG_NAME: &'static str

The name shown for instants of this kind in Debug output.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".