Struct fugit::Duration

pub struct Duration<T, const NOM: u32, const DENOM: u32> { /* private fields */ }

Represents a duration of time.

The generic T can either be u32 or u64, and the const generics represent the ratio of the ticks contained within the duration: duration in seconds = NOM / DENOM * ticks

Implementations

impl<const NOM: u32, const DENOM: u32> Duration<u32, NOM, DENOM>

pub const fn from_ticks(ticks: u32) -> Self

Create a Duration from a ticks value.

let _d = Duration::<u32, 1, 1_000>::from_ticks(1);

pub const fn ticks(&self) -> u32

Extract the ticks from a Duration.

let d = Duration::<u32, 1, 1_000>::from_ticks(234);

assert_eq!(d.ticks(), 234);

pub const fn checked_add<const O_NOM: u32, const O_DENOM: u32>(
self,
other: Duration<u32, O_NOM, O_DENOM>
) -> Option<Self>

Add two durations while checking for overflow.

let d1 = Duration::<u32, 1, 1_000>::from_ticks(1);
let d2 = Duration::<u32, 1, 1_000>::from_ticks(2);
let d3 = Duration::<u32, 1, 1_000>::from_ticks(u32::MAX);

assert_eq!(d1.checked_add(d2).unwrap().ticks(), 3);
assert_eq!(d1.checked_add(d3), None);

pub const fn checked_sub<const O_NOM: u32, const O_DENOM: u32>(
self,
other: Duration<u32, O_NOM, O_DENOM>
) -> Option<Self>

Subtract two durations while checking for overflow.

let d1 = Duration::<u32, 1, 1_000>::from_ticks(1);
let d2 = Duration::<u32, 1, 1_000>::from_ticks(2);
let d3 = Duration::<u32, 1, 1_000>::from_ticks(u32::MAX);

assert_eq!(d2.checked_sub(d1).unwrap().ticks(), 1);
assert_eq!(d1.checked_sub(d3), None);

pub const fn const_partial_cmp<const R_NOM: u32, const R_DENOM: u32>(
self,
other: Duration<u32, R_NOM, R_DENOM>
) -> Option<Ordering>

Const partial comparison.

let d1 = Duration::<u32, 1, 1_00>::from_ticks(1);
let d2 = Duration::<u32, 1, 1_000>::from_ticks(1);

assert_eq!(d1.const_partial_cmp(d2), Some(core::cmp::Ordering::Greater));

pub const fn const_eq<const R_NOM: u32, const R_DENOM: u32>(
self,
other: Duration<u32, R_NOM, R_DENOM>
) -> bool

Const equality check.

let d1 = Duration::<u32, 1, 1_00>::from_ticks(1);
let d2 = Duration::<u32, 1, 1_000>::from_ticks(10);

assert!(d1.const_eq(d2));

pub const fn const_try_into<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Option<Duration<u32, O_NOM, O_DENOM>>

Const try into, checking for overflow.

let d1 = Duration::<u32, 1, 1_00>::from_ticks(1);
let d2: Option<Duration::<u32, 1, 1_000>> = d1.const_try_into();

assert_eq!(d2.unwrap().ticks(), 10);

pub const fn try_into_rate<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Option<Rate<u32, O_NOM, O_DENOM>>

Const try into rate, checking for divide-by-zero.

let d1 = Duration::<u32, 1, 1_000>::from_ticks(2);
let r1: Option<Rate::<u32, 1, 1>> = d1.try_into_rate();

assert_eq!(r1.unwrap().raw(), 500);

pub const fn into_rate<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Rate<u32, O_NOM, O_DENOM>

Convert from duration to rate.

pub const fn convert<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Duration<u32, O_NOM, O_DENOM>

Convert between bases for a duration.

Unfortunately not a From impl due to collision with the std lib.

let d1 = Duration::<u32, 1, 100>::from_ticks(1);
let d2: Duration::<u32, 1, 1_000> = d1.convert();

assert_eq!(d2.ticks(), 10);

Can be used in const contexts. Compilation will fail if the conversion causes overflow

const TICKS: u32= u32::MAX - 10;
const D1: Duration::<u32, 1, 100> = Duration::<u32, 1, 100>::from_ticks(TICKS);
// Fails conversion due to tick overflow
const D2: Duration::<u32, 1, 200> = D1.convert();

pub const fn to_nanos(&self) -> u32

Convert the Duration to an integer number of nanoseconds.

pub const fn to_micros(&self) -> u32

Convert the Duration to an integer number of microseconds.

pub const fn to_millis(&self) -> u32

Convert the Duration to an integer number of milliseconds.

pub const fn to_secs(&self) -> u32

Convert the Duration to an integer number of seconds.

pub const fn to_minutes(&self) -> u32

Convert the Duration to an integer number of minutes.

pub const fn to_hours(&self) -> u32

Convert the Duration to an integer number of hours.

pub const fn nanos(val: u32) -> Duration<u32, NOM, DENOM>

Shorthand for creating a duration which represents nanoseconds.

pub const fn micros(val: u32) -> Duration<u32, NOM, DENOM>

Shorthand for creating a duration which represents microseconds.

pub const fn millis(val: u32) -> Duration<u32, NOM, DENOM>

Shorthand for creating a duration which represents milliseconds.

pub const fn secs(val: u32) -> Duration<u32, NOM, DENOM>

Shorthand for creating a duration which represents seconds.

pub const fn minutes(val: u32) -> Duration<u32, NOM, DENOM>

Shorthand for creating a duration which represents minutes.

pub const fn hours(val: u32) -> Duration<u32, NOM, DENOM>

Shorthand for creating a duration which represents hours.

impl<const NOM: u32, const DENOM: u32> Duration<u64, NOM, DENOM>

pub const fn from_ticks(ticks: u64) -> Self

Create a Duration from a ticks value.

let _d = Duration::<u64, 1, 1_000>::from_ticks(1);

pub const fn ticks(&self) -> u64

Extract the ticks from a Duration.

let d = Duration::<u64, 1, 1_000>::from_ticks(234);

assert_eq!(d.ticks(), 234);

pub const fn checked_add<const O_NOM: u32, const O_DENOM: u32>(
self,
other: Duration<u64, O_NOM, O_DENOM>
) -> Option<Self>

Add two durations while checking for overflow.

let d1 = Duration::<u64, 1, 1_000>::from_ticks(1);
let d2 = Duration::<u64, 1, 1_000>::from_ticks(2);
let d3 = Duration::<u64, 1, 1_000>::from_ticks(u64::MAX);

assert_eq!(d1.checked_add(d2).unwrap().ticks(), 3);
assert_eq!(d1.checked_add(d3), None);

pub const fn checked_sub<const O_NOM: u32, const O_DENOM: u32>(
self,
other: Duration<u64, O_NOM, O_DENOM>
) -> Option<Self>

Subtract two durations while checking for overflow.

let d1 = Duration::<u64, 1, 1_000>::from_ticks(1);
let d2 = Duration::<u64, 1, 1_000>::from_ticks(2);
let d3 = Duration::<u64, 1, 1_000>::from_ticks(u64::MAX);

assert_eq!(d2.checked_sub(d1).unwrap().ticks(), 1);
assert_eq!(d1.checked_sub(d3), None);

pub const fn const_partial_cmp<const R_NOM: u32, const R_DENOM: u32>(
self,
other: Duration<u64, R_NOM, R_DENOM>
) -> Option<Ordering>

Const partial comparison.

let d1 = Duration::<u64, 1, 1_00>::from_ticks(1);
let d2 = Duration::<u64, 1, 1_000>::from_ticks(1);

assert_eq!(d1.const_partial_cmp(d2), Some(core::cmp::Ordering::Greater));

pub const fn const_eq<const R_NOM: u32, const R_DENOM: u32>(
self,
other: Duration<u64, R_NOM, R_DENOM>
) -> bool

Const equality check.

let d1 = Duration::<u64, 1, 1_00>::from_ticks(1);
let d2 = Duration::<u64, 1, 1_000>::from_ticks(10);

assert!(d1.const_eq(d2));

pub const fn const_try_into<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Option<Duration<u64, O_NOM, O_DENOM>>

Const try into, checking for overflow.

let d1 = Duration::<u64, 1, 1_00>::from_ticks(1);
let d2: Option<Duration::<u64, 1, 1_000>> = d1.const_try_into();

assert_eq!(d2.unwrap().ticks(), 10);

pub const fn try_into_rate<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Option<Rate<u64, O_NOM, O_DENOM>>

Const try into rate, checking for divide-by-zero.

let d1 = Duration::<u64, 1, 1_000>::from_ticks(2);
let r1: Option<Rate::<u64, 1, 1>> = d1.try_into_rate();

assert_eq!(r1.unwrap().raw(), 500);

pub const fn into_rate<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Rate<u64, O_NOM, O_DENOM>

Convert from duration to rate.

pub const fn convert<const O_NOM: u32, const O_DENOM: u32>(
self
) -> Duration<u64, O_NOM, O_DENOM>

Convert between bases for a duration.

Unfortunately not a From impl due to collision with the std lib.

let d1 = Duration::<u64, 1, 100>::from_ticks(1);
let d2: Duration::<u64, 1, 1_000> = d1.convert();

assert_eq!(d2.ticks(), 10);

Can be used in const contexts. Compilation will fail if the conversion causes overflow

const TICKS: u64= u64::MAX - 10;
const D1: Duration::<u64, 1, 100> = Duration::<u64, 1, 100>::from_ticks(TICKS);
// Fails conversion due to tick overflow
const D2: Duration::<u64, 1, 200> = D1.convert();

pub const fn to_nanos(&self) -> u64

Convert the Duration to an integer number of nanoseconds.

pub const fn to_micros(&self) -> u64

Convert the Duration to an integer number of microseconds.

pub const fn to_millis(&self) -> u64

Convert the Duration to an integer number of milliseconds.

pub const fn to_secs(&self) -> u64

Convert the Duration to an integer number of seconds.

pub const fn to_minutes(&self) -> u64

Convert the Duration to an integer number of minutes.

pub const fn to_hours(&self) -> u64

Convert the Duration to an integer number of hours.

pub const fn nanos(val: u64) -> Duration<u64, NOM, DENOM>

Shorthand for creating a duration which represents nanoseconds.

pub const fn micros(val: u64) -> Duration<u64, NOM, DENOM>

Shorthand for creating a duration which represents microseconds.

pub const fn millis(val: u64) -> Duration<u64, NOM, DENOM>

Shorthand for creating a duration which represents milliseconds.

pub const fn secs(val: u64) -> Duration<u64, NOM, DENOM>

Shorthand for creating a duration which represents seconds.

pub const fn minutes(val: u64) -> Duration<u64, NOM, DENOM>

Shorthand for creating a duration which represents minutes.

pub const fn hours(val: u64) -> Duration<u64, NOM, DENOM>

Shorthand for creating a duration which represents hours.