Struct fugit::Instant

pub struct Instant<T, const NOM: u32, const DENOM: u32> { /* private fields */ }

Represents an instant in time.

The generic T can either be u32 or u64, and the const generics represent the ratio of the ticks contained within the instant: instant in seconds = NOM / DENOM * ticks

Implementations

impl<const NOM: u32, const DENOM: u32> Instant<u32, NOM, DENOM>

pub const fn from_ticks(ticks: u32) -> Self

Create an Instant from a ticks value.

let _i = Instant::<u32, 1, 1_000>::from_ticks(1);

pub const fn ticks(&self) -> u32

Extract the ticks from an Instant.

let i = Instant::<u32, 1, 1_000>::from_ticks(234);

assert_eq!(i.ticks(), 234);

pub const fn const_cmp(self, other: Self) -> Ordering

Const comparison of Instants.

let i1 = Instant::<u32, 1, 1_000>::from_ticks(1);
let i2 = Instant::<u32, 1, 1_000>::from_ticks(2);

assert_eq!(i1.const_cmp(i2), core::cmp::Ordering::Less);

This function takes into account that ticks might wrap around. If the absolute values of self and other differ by more than half the possible range, it is assumed that an overflow occured and the result is reversed:

let i1 = Instant::<u32, 1, 1_000>::from_ticks(u32::MAX);
let i2 = Instant::<u32, 1, 1_000>::from_ticks(1);

assert_eq!(i1.const_cmp(i2), core::cmp::Ordering::Less);

pub const fn duration_since_epoch(self) -> Duration<u32, NOM, DENOM>

Duration between since the start of the Instant. This assumes an instant which won’t wrap within the execution of the program.

let i = Instant::<u32, 1, 1_000>::from_ticks(11);

assert_eq!(i.duration_since_epoch().ticks(), 11);

pub const fn checked_duration_since( self, other: Self ) -> Option<Duration<u32, NOM, DENOM>>

Duration between Instants.

let i1 = Instant::<u32, 1, 1_000>::from_ticks(1);
let i2 = Instant::<u32, 1, 1_000>::from_ticks(2);

assert_eq!(i1.checked_duration_since(i2), None);
assert_eq!(i2.checked_duration_since(i1).unwrap().ticks(), 1);

pub const fn checked_sub_duration<const O_NOM: u32, const O_DENOM: u32>( self, other: Duration<u32, O_NOM, O_DENOM> ) -> Option<Self>

Subtract a Duration from an Instant while checking for overflow.

let i = Instant::<u32, 1, 1_000>::from_ticks(1);
let d = Duration::<u32, 1, 1_000>::from_ticks(1);

assert_eq!(i.checked_sub_duration(d).unwrap().ticks(), 0);

pub const fn checked_add_duration<const O_NOM: u32, const O_DENOM: u32>( self, other: Duration<u32, O_NOM, O_DENOM> ) -> Option<Self>

Add a Duration to an Instant while checking for overflow.

let i = Instant::<u32, 1, 1_000>::from_ticks(1);
let d = Duration::<u32, 1, 1_000>::from_ticks(1);

assert_eq!(i.checked_add_duration(d).unwrap().ticks(), 2);

impl<const NOM: u32, const DENOM: u32> Instant<u64, NOM, DENOM>

pub const fn from_ticks(ticks: u64) -> Self

Create an Instant from a ticks value.

let _i = Instant::<u64, 1, 1_000>::from_ticks(1);

pub const fn ticks(&self) -> u64

Extract the ticks from an Instant.

let i = Instant::<u64, 1, 1_000>::from_ticks(234);

assert_eq!(i.ticks(), 234);

pub const fn const_cmp(self, other: Self) -> Ordering

Const comparison of Instants.

let i1 = Instant::<u64, 1, 1_000>::from_ticks(1);
let i2 = Instant::<u64, 1, 1_000>::from_ticks(2);

assert_eq!(i1.const_cmp(i2), core::cmp::Ordering::Less);

This function takes into account that ticks might wrap around. If the absolute values of self and other differ by more than half the possible range, it is assumed that an overflow occured and the result is reversed:

let i1 = Instant::<u64, 1, 1_000>::from_ticks(u64::MAX);
let i2 = Instant::<u64, 1, 1_000>::from_ticks(1);

assert_eq!(i1.const_cmp(i2), core::cmp::Ordering::Less);

pub const fn duration_since_epoch(self) -> Duration<u64, NOM, DENOM>

Duration between since the start of the Instant. This assumes an instant which won’t wrap within the execution of the program.

let i = Instant::<u64, 1, 1_000>::from_ticks(11);

assert_eq!(i.duration_since_epoch().ticks(), 11);

pub const fn checked_duration_since( self, other: Self ) -> Option<Duration<u64, NOM, DENOM>>

Duration between Instants.

let i1 = Instant::<u64, 1, 1_000>::from_ticks(1);
let i2 = Instant::<u64, 1, 1_000>::from_ticks(2);

assert_eq!(i1.checked_duration_since(i2), None);
assert_eq!(i2.checked_duration_since(i1).unwrap().ticks(), 1);

pub const fn checked_sub_duration<const O_NOM: u32, const O_DENOM: u32>( self, other: Duration<u64, O_NOM, O_DENOM> ) -> Option<Self>

Subtract a Duration from an Instant while checking for overflow.

let i = Instant::<u64, 1, 1_000>::from_ticks(1);
let d = Duration::<u64, 1, 1_000>::from_ticks(1);

assert_eq!(i.checked_sub_duration(d).unwrap().ticks(), 0);

pub const fn checked_add_duration<const O_NOM: u32, const O_DENOM: u32>( self, other: Duration<u64, O_NOM, O_DENOM> ) -> Option<Self>

Add a Duration to an Instant while checking for overflow.

let i = Instant::<u64, 1, 1_000>::from_ticks(1);
let d = Duration::<u64, 1, 1_000>::from_ticks(1);

assert_eq!(i.checked_add_duration(d).unwrap().ticks(), 2);